Add simple tutorial/evaluation pdu driver
For tutorial purposes and trivial evaluation installations (e.g.
a laptop) adds a localbase class localbase.py for a non-networked
pdu and a driver class localcmdline.py for a local pdu which is
controlled by the command line. This could either be and Arduino
with relays over USB serial or a RPi over ethernet. The actual
command is dependent on whatever your h/w needs, so it's
commented and replaced with a print in localcmdline.py. Entries
are added in strategies.py and lavapdu.conf.