drivers: Add drivers for Devantech relays
Add drivers for the network-controlled interface of Devantech relays:
http://www.robot-electronics.co.uk/products/relay-modules/ethernet-relay.html

This includes a driver for Devantech ETH002, ETH0621, ETH484, ETH008 and
ETH8020 Ethernet relays which are all based on the same
network-controlled interface.

TCP ports and password-protected TCP sessions are supported. When a TCP
session is password-protected, it logs in when initiating and logs out
automatically when closing the session.

You need at least an "ip" attribute in pdu's configuration and you can
add "port" and "password" attributes if needed.

Signed-off-by: Quentin Schulz <quentin.schulz@free-electrons.com>