Use docker in travis-ci
The native travis environment is based on trusty, which is somewhat
outdated by now. This could be seen by the travis.yml pulling in packages
from xenial to try and get somewhat new enough bits.

Improve this by running the tests in a docker environment so the
test environment can both be more modern and deterministic. As an added
bonus this makes it easier to do test runs of the tests outside of
travis.

Signed-off-by: Sjoerd Simons <sjoerd.simons@collabora.co.uk>