Example udev file to set ykush permissions
If pdudaemon isn't running as root it can't access the ykush USB
devices. Add an example udev rules file to change the owner of the
device nodes to pdudaemon.

A local (static) pdudaemon user needs to be created to work well as
otherwise the pdudaemon user only exists when pdudaemon is already
running.

Signed-off-by: Sjoerd Simons <sjoerd.simons@collabora.co.uk>