Add script for finding EnerGenie device IDs
It turns out that the device ID of the EnerGenie devices isn't the same as
the serial number printed on the unit. This script will return the IDs of
the devices found connected. The ID needs to be used in the "device" field
when configuring pdudaemon.conf to use an energenie PDU.

Signed-off-by: Martyn Welch <martyn.welch@collabora.com>