Add Anel NET-PwrCtrl driver
The Anel NET-PwrCtrl family of PDUs feature 3 to 8 ports that can be
controlled via Ethernet.

I chose the HTTP protocol over the UDP one, as, by default, the UDP
protocol uses privileged ports 77 and 75 and therefore would require
elevated privileges for the daemon. The username and password can
optionally be obfuscated with Base64/XOR; we should definitely _not_
implement this feature and lull the user into a false sense of security.

Testing has only been conducted with NET-PwrCtrl HOME.

Signed-off-by: Stefan Wiehler <stefan.wiehler@missinglinkelectronics.com>