pdudaemon: bcu: Init "remote control" every on/off
After adding a bcu-based board to a boardfarm it works initially but
eventually fails to reboot. It is not clear why settings are sometimes
lost but state is 'held' inside an i2c-to-gpio expander so it's better
to not rely on it.

This can be solved by manually running "bcu init" so fix by making
pdudaemon reinitialize "remote control" mode on every operation, similar
to what "bcu reset" does.

With this patch board reset is still available after a few days.

Signed-off-by: Leonard Crestez <leonard.crestez@nxp.com>