Add driver for Netio 4 PDU
The Netio 4 family of 4 port 230V PDUs consists of various (slightly
different) variants, all offering the same telnet interface to switch
the power outlets. While the Netio 4 family was obsoleted, a successor
is available in form of the Netio PowerBOX 4KF. According to vendor
information, the telnet interface is backwards compatible to the one of
the Netio 4 devices. However, I don't have such a device at hand, thus
marking the driver as Netio 4 compatible only.

Signed-off-by: Julian Haller <julian.haller@philips.com>