Implement ESPHome switch control over HTTP API
ESPHome is an interesting project which allows custom firmware
to be written in YAML for a range of ESP8266/ESP32 connected MCUs.

This project is quite interesting for the pdudaemon project as a
number of PDUs and smart switches use MCUs supported by ESPHome.

To remove the cloud element of some of these PDUs by writing
your own firmware becomes very simple with ESPHome.

ESPHome provides two methods of communication with devices: an
HTTP web API (which is meant as a "simple" debugging interface)
as well as a custom API endpoint which uses protocol buffers.

For now, implement support for turning switch components on/off
over the HTTP web API.

Signed-off-by: Christopher Obbard <chris.obbard@collabora.com>