Add missing delay into request mode reboot command
Currently rebooting a device using the request mode doesn't delay in
between turning the port off and on again. This can cause glitches since
the device probably has some kind of holdup which may not fully drain.
When pdudaemon is in daemon mode, the delay between the off and on state
defaults to 5 seconds, set the same default when rebooting a port in
request mode. While we are here, add an optional argument to allow the
user to tweak the dwell time.

Signed-off-by: Christopher Obbard <chris.obbard@collabora.com>