Add normally open/close relay logic selection to Devantech
In the conf file relay logic can be set using "logic" variable. The
variable can be set to "NO" for normally open. And it can be set to
"NC" for normally closed.

For compatibility, "logic" variable is set to "NO" by default if no
value is provided from the config file. So it will work as the same as
before this commit.

Additionally, an example configuration is added to the
share/pdudaemon.conf file.