esphome: Use entity_id string to identify port rather than a port number
The esphome driver currently requires an array in the configuration called
switch_ids which really is a mapping of port number -> ESPHome entity_id.

In #135 pdudaemon gained support for the ability to directly pass as
string as the port name rather than just numbers. Let's simplify the driver
such that we don't need the mapping table and just allow users to pass the
entity_id as the port name.

This may break existing users of the esphome driver as instead pdudaemon
will now pass the port number directly to esphome rather than the mapping
in the configuration file, but esphome will return a 404 error so that can
easily be detected. Since I am probably the only user of the esphome driver
in pdudaemon, I am happy to mark this as a breaking change.

Fixes: #138
Signed-off-by: Christopher Obbard <chris.obbard@collabora.com>