Device identification for novel Cleware USB Switch devices
Recently, support for the Cleware USB Switch 4 was added. With these
newer devices, the serial number is not stored as an HID serial, but has
to be obtained in another way. As it turns out, novel devices of the USB
Switch 8 type also behave like the USB Switch 4, when considering the
serial number.

This commit updates the method to find the device by serial number, so
that it supports both types of serial number identification for the USB
Switch 4 and 8.

Signed-off-by: Sietze van Buuren <Sietze.vanBuuren@de.bosch.com>