pyproject: limit pysnmp and pyasn1
These are required since pdudaemon still uses the sync hlapi internally
for some of the backends. To remedy this a new kind of backend
implementation needs to be written which supports backends with async
code. Then we need to port existing users over to the new async API.

Limit to versions which support the sync hlapi for now.

Signed-off-by: Rouven Czerwinski <rouven@czerwinskis.de>