Rework pdurunner.do_job to return True on success
Currently do_job returns False if the job failed, but returns the state
of driver.handle if the job succeeded. Since driver.handle is of void
type, this results in do_job returning None if the job succeeded.
Rework do_job to return a boolean such that False means the job succeeded
and True means the job failed. This allows easier mapping for the job
status further up the call chain.

Signed-off-by: Christopher Obbard <christopher.obbard@linaro.org>