TPLink: Don't rely on undocumented user specified port naming for indexing
When attempting to use this driver with a freshly installed device, with
the default port names, "Plug 1", "Plug 2" and "Plug 3", attempts to
turn on a socket fail:

    pdud.pdu.KP303.WARNING Traceback (most recent call last):
      File "/usr/lib/python3/dist-packages/pdudaemon/pdurunner.py", line 56, in do_job
        return self.driver.handle(request, port)
               ^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^
      File "/usr/lib/python3/dist-packages/pdudaemon/drivers/driver.py", line 70, in handle
        self.port_on(port_number)
      File "/usr/lib/python3/dist-packages/pdudaemon/drivers/driver.py", line 81, in port_on
        self.port_interaction("on", port_number)
      File "/usr/lib/python3/dist-packages/pdudaemon/drivers/tplink.py", line 110, in port_interaction
        context = self.get_context(port_number)
                  ^^^^^^^^^^^^^^^^^^^^^^^^^^^^^
      File "/usr/lib/python3/dist-packages/pdudaemon/drivers/tplink.py", line 97, in get_context
        child_port = (int(child['alias'].split("_")[-1]) + 1)
                      ^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^
    ValueError: invalid literal for int() with base 10: 'Plug 1'

It seems that the driver is expecting the ports to be named in a
specific format, which isn't documented and I'd argue isn't warranted.

Modify the driver to enable the ports to be identified either by the
port name or by port index. This way the following options are available
to turn on the first socket in the default case:

  $ curl "http://0.0.0.0:16421/power/control/on?hostname=KP303&port=1"
  $ curl "http://0.0.0.0:16421/power/control/on?hostname=KP303&port=Plug%201"

With custom names for the ports also working for example with a port
named "Test":
  $ curl "http://0.0.0.0:16421/power/control/on?hostname=KP303&port=Test"

Additionally, raise errors when port can't be found or device isn't
initialised.

Signed-off-by: Martyn Welch <martyn.welch@collabora.com>