Add support for different firmware versions of APC AP7900
The APC AP7900 has different UIs depending on its firmware.
The current code only works with the UI provided in version 2.7.3 (or
other versions). This adds processing confirmed in version 3.7.3.

Processing based on firmware version is achieved by setting the
`firmware_version` variable in the `config` file's `pdus` section for apcbase
driver, which is used to switch processing for each version.

Additionally, since this has not been verified for all firmware versions, it
reflects the processing for firmware versions confirmed in 2.7.3 and 3.7.3.

Signed-off-by: Nobuhiro Iwamatsu <iwamatsu@nigauri.org>